backend/requirements.txt
fastapi
uvicorn
python-dotenv
mysql-connector-python
torch
torchvision
pillow
opencv-python
numpy
tensorflow
